web/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
python-jose[cryptography]
bcrypt
jinja2
httpx
python-multipart